pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "antipublicbot"
version = "0.1.0"
description = "High-performance Telegram anti-public bot for very large text datasets"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0,<22.0",
  "python-dotenv>=1.0.0"
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]
